fix: 장바구니 삭제시 상품의 id값으로 삭제

+++ app/src/main/java/woowacourse/shopping/ui/cart/CartViewModel.kt
@@ -23,9 +23,8 @@ class CartViewModel(
         _cartProducts.value = cartRepository.cartProducts().toProducts()
     }
 
-    fun deleteCartProduct(id: Int) = viewModelScope.launch {
-        cartRepository.deleteCartProduct(id.toLong())
-
+    fun deleteCartProduct(id: Long) = viewModelScope.launch {
+        cartRepository.deleteCartProduct(id)
         _onCartProductDeleted.value = true
     }
 }
